Added k-d tree for neighbor search and changed xy-coordinates to struct

+++ include/parameters_xml.hpp
@@ -15,7 +15,7 @@ struct parameters
 {
 	int         N;
     float       DT;
-    int         TIME;
+    float       TIME;
     bool        SPATIAL;
     int         LENGTH;
     int         WIDTH;
